test: fix phase6VectorStatistics type witness (AVG returns Double, not Float)
AVG() returns a Double (issue #4518), so result.<Float>getProperty(...) on the
AVG-wrapped avg_magnitude/sparsity_pct columns threw a ClassCastException. Use the
matching <Double> witness.

+++ engine/src/test/java/com/arcadedb/query/sql/functions/vector/SQLVectorDatabaseFunctionsTest.java
@@ -525,8 +525,9 @@ class SQLVectorDatabaseFunctionsTest extends TestHelper {
 
       assertThat(rs.hasNext()).isTrue();
       final Result result = rs.next();
-      assertThat(result.<Float>getProperty("avg_magnitude")).isNotNull();
-      assertThat(result.<Float>getProperty("sparsity_pct")).isNotNull();
+      // AVG() always returns a Double (issue #4518), so use the matching type witness here.
+      assertThat(result.<Double>getProperty("avg_magnitude")).isNotNull();
+      assertThat(result.<Double>getProperty("sparsity_pct")).isNotNull();
     });
   }
